Enable scummvmsa on S922X
scummvmsa has been disabled on S922X a long time ago (it was already in jelos). It was compilable but did not start. The bad thing is that retroarch scummvm uses a script that is part of scummvmsa. So in the end scummvm is completely unusable on the S922X.
I don't know what the problem was but it doesn't exist any more. scummvmsa is compilable and working for S922X and S922X-PANFROST.