Write `bluetooth-agent` pid on process start.
I've experienced the `Pair a Bluetooth device manually` failing to list
any devices.

Upon inspection, `/run/bluetooth-agent.pid` is not getting the PID of
the agent written into it, making `do_devlist` in `rocknix-bluetooth`
fail.

With this change, the pid will be written out by the unit that starts
bluetooth-agent.

Tested on a Retroid Pocket 5 by writing the modified unit to
/storage/.config/system.d/bluetooth-agent.service and enabling it.

On reboot and after toggling bluetooth from disabled to enabled
bluetooth devices are now listed.