Write `bluetooth-agent` pid on process start.
I've experienced the `Pair a Bluetooth device manually` failing to list
any devices.

Upon inspection, `/run/bluetooth-agent.pid` is not getting the PID of
the agent written into it, making `do_devlist` in `rocknix-bluetooth`
fail.

With this change, the pid will be written out by
`rocknix-bluetooth-agent` on start.

I have also cleaned up the call to pgrep in `rocknix-bluetooth` to
ensure that the PID is being set only by the agent.

Tested on a Retroid Pocket 5 by writing bind mounting the two modified
files and starting / stopping bluetooth.